% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{app}
\alias{app}
\title{app}
\usage{
app(
  shiny_args = list(launch.browser = F),
  render_args = NULL,
  indir = getwd(),
  userdir = getwd()
)
}
\arguments{
\item{shiny_args}{\code{list}, optional list of arguments conveyed to
\code{rmarkdown::run} \code{shiny_args} argument.}

\item{render_args}{\code{list}, optional list of arguments conveyed to
\code{rmarkdown::run} \code{render_args} argument.}

\item{indir}{\code{character(1)}, a location on the server side
containing data directories.}

\item{userdir}{\code{character(1)}, a location on the server side containing project directories.}
}
\value{
Nada.
}
\description{
app
}
\author{
Todor Kondić
}
