% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{attch}
\alias{attch}
\title{Paste With No Separator}
\usage{
attch(...)
}
\arguments{
\item{...}{Strings to paste together.}
}
\value{
Pasted string.
}
\description{
Paste with no separator.
}
\details{
.. content for \details{} ..
}
\author{
Todor Kondić
}
