% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envopts.R
\name{empty_envopts}
\alias{empty_envopts}
\title{Create an Empty \code{envopts} Object}
\usage{
empty_envopts(
  projects = NULL,
  top_data_dir = NULL,
  users_dir = NULL,
  metfrag_db_dir = NULL,
  metfrag_jar = NULL,
  java_bin = NULL,
  metfrag_max_proc = NULL,
  no_structure_plots = NULL
)
}
\arguments{
\item{projects}{\code{character(1)}, a directory which contains all
shinyscreen projects directories. A single project directory
contains input and output files.}

\item{top_data_dir}{\code{character(1)}, a directory which contains all
\code{data} directories. A single \code{data} directory contains \code{mzML}
spectrometry data files.}

\item{users_dir}{\code{character(1)}, a location on the server side
containing individual user directories.}

\item{metfrag_db_dir}{\code{character(1)}, a path to the directory which
contains MetFrag databases.}

\item{metfrag_jar}{\code{character(1)}, a path to MetFrag JAR file.}

\item{java_bin}{\code{character(1)}, a path to java runtime. If no path
has been supplied, we will try to detect it.}

\item{metfrag_max_proc}{\code{integer(1)}, maximum number of CPU cores
available for MetFrag. If no number has been supplied, we will
try to detect the number of logical CPUs and go with that.}

\item{no_structure_plots}{\code{logical(1)}, if T, structures will not
be plotted, even when it is possibile.}
}
\value{
An \code{envopts} object.
}
\description{
Create an Empty \code{envopts} Object
}
\details{
An \code{envopts} object is Shinyscreen way to store settings
related to a specific computing environment. Information such
as the run time path to a MetFrag JAR will vary from one to
another setup and we need to convey this to the \code{shinyscreen}
pipeline.
}
\author{
Todor Kondić
}
