% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{gen_file_table}
\alias{gen_file_table}
\title{Generate and Load the RMassBank Settings File}
\usage{
gen_file_table(fn_data, n_cmpd, wd)
}
\arguments{
\item{fn_data}{The mzML filename.}

\item{n_cmpd}{Number of compounds.}

\item{wd}{Directory under which results are archived.}
}
\value{
File path of the file table.
}
\description{
Generates file table.
}
\author{
Todor Kondić
}
