% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{gen_ftable}
\alias{gen_ftable}
\title{Generate and Load the RMassBank Settings File}
\usage{
gen_ftable(fn_data, wd, n_cmpd)
}
\arguments{
\item{fn_data}{The mzML filename.}

\item{wd}{Directory under which results are archived.}

\item{n_cmpd}{Number of compounds.}
}
\value{
File path of the file table.
}
\description{
Generates file table.
}
\author{
Todor Kondić
}
