% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{mb.do}
\alias{mb.do}
\title{Perform the Mass Bank workflow}
\usage{
mb.do(mb, rdir = ".")
}
\arguments{
\item{mb}{The list of prepared mbWorkspace objects.}

\item{rdir}{Root data dir.}
}
\value{
The named list of processed mbWorkspace objects.
}
\description{
Perform the Mass Bank workflow on the prepared mbWorkspace
objects. FOR THE RECORD: writing documentation is boring. Anyway,
record results in the info subfolder of each data subdir.
}
\details{
.. content for \details{} ..
}
\author{
Todor Kondić
}
