% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{mb.p}
\alias{mb.p}
\title{Parallel Mass Bank Workflow}
\usage{
mb.p(mb, infodir, fn_stgs, cl = F)
}
\arguments{
\item{mb}{List of mass bank workflow objects}

\item{infodir}{List of subdirs containing info lists.}

\item{fn_stgs}{List of settings files.}

\item{cl}{Cluster.}
}
\value{
A named list of mbWorkspace objects. The names are derived
from the input mb sequence.
}
\description{
Interface to parallelised Mass Bank workflow.
}
\author{
Todor Kondić
}
