% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{mb.prep}
\alias{mb.prep}
\title{Prepare mbWorkspace objects}
\usage{
mb.prep(w)
}
\arguments{
\item{w}{A list of spectral workspace inputs.}
}
\value{
Named list of prepared mbWorkspace objects.
}
\description{
Creates and prepares mbWorkspace objects before the full workflow
is performed on them. In process, create directory \code{info} as a
subdir of any particular data dir and place the starting info list
there.
}
\author{
Todor Kondić
}
