% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{mb.prep.single}
\alias{mb.prep.single}
\title{Prepare Single mbWorkspace object}
\usage{
mb.prep.single(w, fn_info, fn_stgs)
}
\arguments{
\item{w}{MsmsWorkspace object.}

\item{fn_info}{Filename of the infolist to be generated.}

\item{fn_stgs}{Filename of the RMassBank settings.}
}
\value{
A mbWorkspace object.
}
\description{
Prepare single mbWorkspace object based on the workspace, the
infolist name and RMassBank settings.
}
\author{
Todor Kondić
}
