% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{mb.prep.v}
\alias{mb.prep.v}
\title{Vectorized mb.prep function.}
\usage{
mb.prep.v(w, fn_info, fn_stgs)
}
\arguments{
\item{w}{A sequence of msmsWorkspaces.}

\item{fn_info}{A sequence of infolist filenams to be generated.}

\item{fn_stgs}{A sequence of settings associated with each
msmsWorkspace object.}
}
\value{
A list of mbWorkspaces.
}
\description{
Vectorize mb.prep function.
}
\author{
Todor Kondić
}
