% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{mb.single}
\alias{mb.single}
\title{Single MassBank workflow.}
\usage{
mb.single(mb, infodir, fn_stgs)
}
\arguments{
\item{mb}{A mbWorkspace object.}

\item{infodir}{Directory containing the infolist.}

\item{fn_stgs}{The settings associated with the mbWorkspace
object.}
}
\value{
A mbWorkflow object.
}
\description{
Performs a single MassBank workflow after preparation.
}
\author{
Todor Kondić
}
