% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{mb.v}
\alias{mb.v}
\title{Vectorised Mass Bank Workflow}
\usage{
mb.v(mb, infodir, fn_stgs)
}
\arguments{
\item{mb}{List of mass bank workflow objects}

\item{infodir}{List of subdirs containing info lists.}

\item{fn_stgs}{List of settings files.}
}
\value{
A named list of mbWorkspace objects. The names are derived
from the input mb sequence.
}
\description{
Interface to vectorised Mass Bank workflow.
}
\details{
.. content for \details{} ..
}
\author{
Todor Kondić
}
