% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{mk_combine_file}
\alias{mk_combine_file}
\title{Combine RMB Settings With Different Collisional Energies}
\usage{
mk_combine_file(sett_fns, fname)
}
\arguments{
\item{sett_fns}{A list of settings files.}

\item{fname}{The name of the combined file.}
}
\value{
fname
}
\description{
Combine RMB settings with different collisional energies into one
settings file with multiple collisional energy entries.
}
\details{
.. content for \details{} ..
}
\author{
Todor Kondić
}
