% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{mk_sett_file}
\alias{mk_sett_file}
\title{Generate RMassBank settings file.}
\usage{
mk_sett_file(sett_alist, file)
}
\arguments{
\item{sett_alist}{The named list of settings that are different
from the RMassBank defaults.}

\item{file}{The name of the YAML specification that will be merged
with the template Rmb settings file.}
}
\description{
Produce the Rmb Settings file
}
\details{
Produce the Rmb Settings file based on the customisation file in
YAML format.
}
\author{
Todor Kondić
}
