% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{new_project}
\alias{new_project}
\title{Create a New Project}
\usage{
new_project(project, datatab = NULL, conf = NULL)
}
\arguments{
\item{project}{\code{character(1)}, path to a directory containing all the projects}

\item{datatab}{\code{datatab}, a \code{data.table} describing data files, optional}

\item{conf}{\code{conf}, a configuration object, optional}
}
\description{
Create a New Project
}
\author{
Todor Kondić
}
