% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{new_runtime_state}
\alias{new_runtime_state}
\title{Create Runtime Configuration}
\usage{
new_runtime_state(project, conf = NULL)
}
\arguments{
\item{project}{\code{character(1)}, project title (and the directory name under \code{projects})}

\item{conf}{\code{conf}, configuration object, optional}
}
\value{
\code{run} substate
}
\description{
Create Runtime Configuration
}
\details{
This is a part of the configuration that can only be
diagnosed at runtime.
}
\author{
Todor Kondić
}
