% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{new_state}
\alias{new_state}
\title{Create Initial State}
\usage{
new_state()
}
\value{
A state object.
}
\description{
Create Initial State
}
\details{
Creates an initial, bare state.
}
\author{
Todor Kondić
}
