% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{no_drama_mkdir}
\alias{no_drama_mkdir}
\title{Create directories without drama}
\usage{
no_drama_mkdir(path)
}
\arguments{
\item{path}{Name of the directory.}
}
\value{
The character string containing the input argument \code{path}.
}
\description{
Create directories without drama.
}
\details{
Create directories without drama.
}
\author{
Todor Kondić
}
