% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{presc.do}
\alias{presc.do}
\title{Prescreening on bunch of files.}
\usage{
presc.do(fn_data, fn_cmpd_l, mode, dest = ".", proc = F, ...)
}
\arguments{
\item{fn_data}{The mzML files. Basis for the out directory name
generation.}

\item{fn_cmpd_l}{The compound list.}

\item{mode}{RMB mode.}

\item{dest}{Destination directory.}

\item{proc}{Amount of processors, or FALSE.}

\item{fn_cmpd_list}{The compound list CSV.}
}
\value{
Nothing useful.
}
\description{
Do the prescreening.
}
\author{
Todor Kondić
}
