% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{presc.shiny}
\alias{presc.shiny}
\title{Prescreening with Shiny}
\usage{
presc.shiny(prescdf, mode, fn_cmpd_l, pal = "Dark2", cex = 0.75,
  rt_digits = 2, m_digits = 4)
}
\arguments{
\item{prescdf}{File table data-frame. Columns: Files,ID,wd,tag,set_name ...}

\item{mode}{RMassBank mode.}

\item{fn_cmpd_l}{Compound list file name.}

\item{pal}{ColorBrewer palette.}

\item{cex}{Size of fonts.}

\item{rt_digits}{Number of decimal places for the retention time.}

\item{m_digits}{Number of decimal places for the mass.}
}
\value{
Nothing useful.
}
\description{
Prescreening using shiny interface.
}
\author{
Jessy Krier

Mira Narayanan

Hiba Mohammed Taha

Anjana Elapavalore

Todor Kondić
}
