% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{presc.single}
\alias{presc.single}
\title{Wrapper for RMB_EIC_Prescreen}
\usage{
presc.single(fn_data, stgs_alist, wd, mode, fn_cmpd_l, ppm_lim_fine = 10,
  EIC_limit = 0.001)
}
\arguments{
\item{fn_data}{The mzML filename.}

\item{stgs_alist}{Settings named list, or a settings filename.}

\item{wd}{Directory under which results are archived.}

\item{mode}{RMB mode.}

\item{fn_cmpd_l}{Filename of the compound list.}

\item{ppm_lim_fine}{The ppm_limit_fine argument to RMB_EIC_Prescreen}

\item{EIC_limit}{Passed down to RMB_EIC_Prescreen.}
}
\value{
result of RMB_EIC_Prescreen
}
\description{
Wrapper for a single prescreening call. Produces output in the
usual mix method places.
}
\author{
Todor Kondić
}
