% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{rendersmiles2}
\alias{rendersmiles2}
\title{Turn SMILES to an Image Using Online Resource}
\usage{
rendersmiles2(smiles, style = "cow", ...)
}
\arguments{
\item{smiles}{The SMILES string.}

\item{style}{Structure style.}

\item{...}{Hand over to renderurl.}
}
\value{
Nothing useful.
}
\description{
Render smiles from an online resource.
}
\author{
Todor Kondić
}
