% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{renderurl}
\alias{renderurl}
\title{Render Compound from an Online Resource}
\usage{
renderurl(depictURL, coords = c(0, 0, 100, 100),
  filename = tempfile(fileext = ".svg"))
}
\arguments{
\item{depictURL}{The URL of the object to plot.}

\item{coords}{The positioning of the image (in data coords).}

\item{filename}{Temp filename.}
}
\value{
Nothing useful.
}
\description{
Helper function for rendersmiles2
}
\author{
Todor Kondić
}
