% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{run}
\alias{run}
\title{Run Shinyscreen Pipeline}
\usage{
run(envopts, project = "", m = NULL, phases = NULL, help = F)
}
\arguments{
\item{envopts}{\code{envopts}, an \code{envopts} object.}

\item{project}{\code{character(1)}, a directory containing input data.}

\item{m}{\code{state}, a Shinyscreen state.}

\item{phases}{\code{character(n)}, a character vector of Shinyscreen
phases.}

\item{help}{\code{logical(1)}, print help?}
}
\description{
Run Shinyscreen Pipeline
}
\details{
TODO.
}
