% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{single.sw}
\alias{single.sw}
\title{RMassBank Spectral Workflow on a Single Compound Mixture}
\usage{
single.sw(fn_data, stgs_alist, wd, fn_cmpd_list, mode,
  readMethod = "mzR", archdir = "archive", lastStep = 8)
}
\arguments{
\item{fn_data}{A mzML data file.}

\item{stgs_alist}{RMassBank settings. It can either be a named
list of settings, or a filename of a YAML file.}

\item{wd}{The name of the work directory.}

\item{fn_cmpd_list}{The file name of he compound list
corresponding to \code{fn_data}.}

\item{mode}{Modes as described in the standard workflow vignette
of RMassBank.}

\item{readMethod}{Default read method is "mzR". Consult the
documentation of \code{msmsRead} for details.}

\item{archdir}{The directory to store R objects created during
workflow execution.}

\item{lastStep}{The last step in the workflow. Default is eight.}
}
\value{
MsmsWorkspace object.
}
\description{
Runs a compound mixture workflow on a single mzML file.
}
\author{
Todor Kondić
}
