% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{sort_spectra}
\alias{sort_spectra}
\title{Sort the Summary Table}
\usage{
sort_spectra(m)
}
\arguments{
\item{m}{}
}
\value{
m
}
\description{
Sets the key specified by DEF_KEY_SUMM and adds second indices,
either from DEF_INDEX_SUMM, or user-specified in
conf[\link{"summary table"}]$order. The order entry is a list of
strings with names of columns in summ, optionally prefixed with a
minus(-) sign. Columns prefixed with the minus are going to be in
ascending order.
}
\author{
Todor Kondić
}
