% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{subset_summary}
\alias{subset_summary}
\title{Subset the Summary Table}
\usage{
subset_summary(m)
}
\arguments{
\item{m}{}
}
\value{
m
}
\description{
Subsets the summary table by applying conditions set out in the
filter subkey of summary table key of the config. Each member of
filter is an expression that and all of them are chained together
using AND logical operation and applied to the summary table.
}
\author{
Todor Kondić
}
