% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{sw.do}
\alias{sw.do}
\title{Perform MassBank Workflow on Multiple Compound Mixtures}
\usage{
sw.do(fn_data, fn_cmpd_list, mode, dest = ".", combine = F, proc = F)
}
\arguments{
\item{fn_data}{List of mzML data filenames to be processed.}

\item{fn_cmpd_list}{Compound list.}

\item{mode}{as in msmsRead.}

\item{dest}{The destination data directory.}

\item{combine}{If TRUE, use combineMultiplicies to merge
workspaces corresponding to different collisional energies.}

\item{proc}{Split work between this amount of processes. If FALSE
(or, 1), run sequential.}
}
\value{
A named list of msmsWorkspace objects.
}
\description{
Performs massbank workflow on multiple mzML files:
}
\details{
\enumerate{
\item Create rdir if it does not exist.
\item Generate data subdirectories inside \code{rdir} that will be used to
keep the processed data belonging to a particular mzML file in
the fileset.
\item Generate the full RMassBank settings files by merging the
sections from the \code{fn_cmpd_list} with the defaults and place those
new settings files in the appropriate data subdirs.
\item Finally perform the spectral workflow on each data file and
place the results in the data subdirs.
}
}
\author{
Todor Kondić
}
