% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{sw.single.1}
\alias{sw.single.1}
\title{RMassBank Spectral Workflow on a Single Compound Mixture (step 1)}
\usage{
sw.single.1(fn_data, stgs_alist, wd, fn_cmpd_list, mode,
  readMethod = "mzR")
}
\arguments{
\item{fn_data}{A mzML data file.}

\item{stgs_alist}{RMassBank settings. It can either be a named
list of settings, or a filename of a YAML file.}

\item{wd}{The name of the work directory.}

\item{fn_cmpd_list}{The file name of he compound list
corresponding to \code{fn_data}.}

\item{mode}{Modes as described in the standard workflow vignette
of RMassBank.}

\item{readMethod}{Default read method is "mzR". Consult the
documentation of \code{msmsRead} for details.}
}
\value{
MsmsWorkspace object.
}
\description{
Runs the first step of the compound mixture workflow on a single mzML file.
}
\author{
Todor Kondić
}
