% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{sw.single.next}
\alias{sw.single.next}
\title{RMassBank Spectral Workflow on a Single Compound Mixture (after step 1)}
\usage{
sw.single.next(w, wd, archdir = "archive", lastStep = 8)
}
\arguments{
\item{w}{The msmsWorkspace object that was processed by
sw.single.1.}

\item{wd}{The current working dir.}

\item{archdir}{The directory to store R objects created during
workflow execution.}

\item{lastStep}{The last step in the workflow. Default is eight.}
}
\value{
MsmsWorkspace object.
}
\description{
Runs steps after sw.single.1 of compound mixture workflow on a
single mzML file.
}
\author{
Todor Kondić
}
