% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{v}
\alias{v}
\title{Vectorised Spectral Workflow.}
\usage{
v(fn_data, stgs_alist, wd, fn_cmpd_list, mode, readMethod = "mzR",
  archdir = "archive")
}
\arguments{
\item{fn_data}{A sequence of mzML input files.}

\item{stgs_alist}{A list of named list of settings, or a list of
filenames of YAML files containing the settings.}

\item{wd}{The list of working directories.}

\item{fn_cmpd_list}{The compound list characterising the mixtures.}

\item{mode}{Same as in msmsRead.}

\item{readMethod}{Same as in msmsRead.}

\item{archdir}{Name of the archive.}
}
\value{
A named list of spectral workspaces. The names are derived
from data filenames.
}
\description{
Interface to vectorised spectral workflow.
}
\author{
Todor Kondić
}
