% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{app}
\alias{app}
\title{app}
\usage{
app(
  projects = getwd(),
  top_data_dir = getwd(),
  metfrag_db_dir = "",
  metfrag_runtime = "",
  shiny_args = list(launch.browser = F),
  render_args = NULL
)
}
\arguments{
\item{projects}{\code{character(1)}, a location on the server side
containing project directories.}

\item{top_data_dir}{\code{character(1)}, a location on the server side
containing data directories.}

\item{metfrag_db_dir}{\code{character(1)}, a location on the server side
containing MetFrag databases.}

\item{metfrag_runtime}{\code{character(1)}, a location on the server side
of the MetFrag jar file.}

\item{shiny_args}{\code{list}, optional list of arguments conveyed to
\code{rmarkdown::run} \code{shiny_args} argument.}

\item{render_args}{\code{list}, optional list of arguments conveyed to
\code{rmarkdown::run} \code{render_args} argument.}
}
\value{
Nada.
}
\description{
app
}
\author{
Todor Kondić
}
