% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{app}
\alias{app}
\title{app}
\usage{
app(envopts, shiny_args = list(launch.browser = F), render_args = NULL)
}
\arguments{
\item{envopts}{\code{envopts}. Shinyscreen environment options.}

\item{shiny_args}{\code{list}, optional list of arguments conveyed to
\code{rmarkdown::run} \code{shiny_args} argument.}

\item{render_args}{\code{list}, optional list of arguments conveyed to
\code{rmarkdown::run} \code{render_args} argument.}

\item{metfrag_runtime}{\code{character(1)}, a location on the server side
of the MetFrag jar file.}
}
\value{
Nada.
}
\description{
app
}
\author{
Todor Kondić
}
