% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envopts.R
\name{envopts}
\alias{envopts}
\title{Create a \code{envopts} Object}
\usage{
envopts(metfrag_db_dir = "", metfrag_jar = "", java_bin = Sys.which("java"))
}
\arguments{
\item{metfrag_db_dir}{\code{character(1)}, a path to the directory which contains MetFrag databases}

\item{metfrag_jar}{\code{character(1)}, a path to MetFrag JAR file}
}
\value{
An \code{envopts} object.
}
\description{
Create a \code{envopts} Object
}
\details{
A \code{envopts} object is Shinyscreen way to store settings
related to a specific computing environment. Information such
as the run time path to a MetFrag JAR will vary from one to
another setup and we need to convey this to the \code{shinyscreen}
pipeline.
}
\author{
Todor Kondić
}
