% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{gen_cmpdl_and_load}
\alias{gen_cmpdl_and_load}
\title{Generate and Load the RMassBank Compound List}
\usage{
gen_cmpdl_and_load(wd, fn_cmpdl)
}
\arguments{
\item{wd}{Directory under which results are archived.}

\item{fn_cmpdl}{The input compound list filename.}
}
\value{
Named list. The key \code{fn_cmpdl} is the path of the
generated compound list and the key \code{n} the number of
compounds.
}
\description{
Generates the RMassBank compound list and loads it.
}
\author{
Todor Kondić
}
