% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{gen_stgs_and_load}
\alias{gen_stgs_and_load}
\title{Generate and Load the RMassBank Settings File}
\usage{
gen_stgs_and_load(fn_data, stgs, wd)
}
\arguments{
\item{fn_data}{The mzML filename.}

\item{wd}{Directory under which results are archived.}

\item{stgs_alist}{Settings named list, or a settings filename.}
}
\value{
result of RMassBank::loadRmbSettings
}
\description{
Generates settings file and loads it.
}
\author{
Todor Kondić
}
