% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{init}
\alias{init}
\title{Initialise Shinyscreen Configuration}
\usage{
init(metfrag_db_dir = "", metfrag_jar = "", java_bin = Sys.which("java"))
}
\arguments{
\item{metfrag_db_dir}{\code{character(1)}, a path to the directory which contains MetFrag databases}

\item{metfrag_jar}{\code{character(1)}, a path to MetFrag JAR file}
}
\value{
Nothing.
}
\description{
Initialise Shinyscreen Configuration
}
\details{
This function is used to inform \code{shinyscreen} about the
working environment. It is only necessary to call it once. The
parameters will be memorised.
}
\author{
Todor Kondić
}
