% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{new_empty_project}
\alias{new_empty_project}
\alias{new_project}
\title{Create a Bare Project}
\usage{
new_empty_project(project, envopts)

new_project(project, envopts, datatab = NULL, conf = NULL)
}
\arguments{
\item{project}{\code{character(1)}, path to a directory containing all
the projects}

\item{envopts}{\code{envopts}, an \code{envopts} object.}

\item{datatab}{\code{datatab}, a \code{data.table} describing data files,
optional}

\item{conf}{\code{conf}, a configuration object, optional}
}
\value{
New project \code{state}
}
\description{
Create a Bare Project

Create a New Fully Defined Project
}
\details{
This function creates a \code{shinyscreen} project with no
configuration. This is not enough to run a pipeline, but will
start things for incremental configuration using the GUI.

After the call to \code{new_project}, the \code{shinyscreen}
pipeline should be ready to start.
}
\seealso{
\code{\link[=new_state]{new_state()}}

\code{\link[=empty_envopts]{empty_envopts()}},
\code{\link[=new_empty_project]{new_empty_project()}}
}
\author{
Todor Kondić
}
