% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{presc.p}
\alias{presc.p}
\title{Parallel version of presc.single}
\usage{
presc.p(cl, fn_data, fn_cmpd_l, mode, ppm_lim_fine = 10,
  EIC_limit = 0.001)
}
\arguments{
\item{cl}{Cluster object.}

\item{fn_data}{Sequence of mzML files.}

\item{fn_cmpd_l}{Filename of the compound list.}

\item{mode}{RMB mode.}

\item{ppm_lim_fine}{See ReSOLUTION.}

\item{EIC_limit}{See ReSOLUTION.}
}
\value{
Nothing useful.
}
\description{
Parallel version of presc.single.
}
\author{
Todor Kondić
}
