% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{presc.plot}
\alias{presc.plot}
\title{Plot the Output of Prescreen}
\usage{
presc.plot(wd, out = "prescreen.pdf", pal = "Accent")
}
\arguments{
\item{wd}{Sequence of data dirs containing the prescreen subdir.}

\item{out}{The name of the output file.}

\item{pal}{ColorBrewer palette name.}
}
\value{
Nothing useful.
}
\description{
Plot the output of prescreen.
}
\author{
Todor Kondić
}
