% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{presc.plot}
\alias{presc.plot}
\title{Plot the Output of Prescreen}
\usage{
presc.plot(prescdf, mode, out = "prescreen.pdf", fn_cmpd_l,
  pal = "Dark2", cex = 0.75, rt_digits = 2, m_digits = 4)
}
\arguments{
\item{prescdf}{File table data-frame. See presc.shiny for details.}

\item{mode}{RMB mode.}

\item{out}{The name of the output file.}

\item{fn_cmpd_l}{The compound list name.}

\item{pal}{ColorBrewer palette name.}

\item{cex}{As in legend.}

\item{rt_digits}{Number of digits after the point for the retention time.}

\item{m_digits}{Number of digits after the point for the mass.}

\item{wd}{Sequence of data dirs containing the prescreen subdir.}

\item{digits}{Number of significant digits for peak ret times.}
}
\value{
Nothing useful.
}
\description{
Plot the output of prescreen.
}
\author{
Todor Kondić

Mira Narayanan

Anjana Elapavalore
}
