% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{serve}
\alias{serve}
\title{serve}
\usage{
serve(indir, topuserdir, user, host = "0.0.0.0", port = 7777)
}
\arguments{
\item{indir}{\code{character(1)}, a location on the server side
containing data directories.}

\item{topuserdir}{\code{character(1)}, a location on the server side
containing individual user directories.}

\item{user}{\code{character(1)}, subdir of topuserdir.}

\item{host}{\code{character(1)}, optional, address where the page is
served.}

\item{port}{\code{integer(1)}, optional, port at which the page is
served.}
}
\value{
Nada.
}
\description{
serve
}
\author{
Todor Kondić
}
