% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{serve}
\alias{serve}
\title{serve}
\usage{
serve(envopts, user, host = "0.0.0.0", port = 7777)
}
\arguments{
\item{envopts}{\code{envopts}, an \code{envopts} object.}

\item{user}{\code{character(1)}, subdir of usersdir.}

\item{host}{\code{character(1)}, optional, address where the page is
served.}

\item{port}{\code{integer(1)}, optional, port at which the page is
served.}

\item{top_data_dir}{\code{character(1)}, a location on the server side
containing data directories.}

\item{metfrag_db_dir}{\code{character(1)}, a location on the server side
containing MetFrag DBs.}
}
\value{
Nada.
}
\description{
serve
}
\author{
Todor Kondić
}
